requirements.txt
asgiref==3.5.1
black==22.3.0
certifi==2020.6.20
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
